+++ scripts/detect_secrets/check.sh
@@ -1,5 +1,6 @@
 #!/bin/sh
 set -euf
 
-git ls-files -z | xargs -0 detect-secrets-hook --baseline gen/secrets.baseline.json
+# Skip files that take a long time and are unlikely to include secrets
+git ls-files -z | grep -zvE '^(sdks/|lib/smithy-output/|infra/helm/|infra/tf/k8s_infra/grafana_dashboards/)' | xargs -0 detect-secrets-hook --baseline gen/secrets.baseline.json