+++ scripts/fern/generate_internal.sh
@@ -1,5 +1,0 @@
-#!/usr/bin/env bash
-set -euf -o pipefail
-
-fern generate --group internal --log-level debug
-