+++ scripts/fern/generate_local.sh
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-set -euf -o pipefail
-
-fern generate --group local_internal --log-level debug
-fern generate --group local_external --log-level debug
-