+++ sdks/typescript/src/api/resources/cloud/resources/games/resources/games/client/requests/index.ts
@@ -1,0 +1,2 @@
+export { GetGamesRequest } from "./GetGamesRequest";
+export { GetGameByIdRequest } from "./GetGameByIdRequest";