+++ sdks/typescript/src/api/resources/cloud/resources/games/resources/matchmaker/client/requests/index.ts
@@ -1,0 +1,1 @@
+export { GetLobbyLogsRequest } from "./GetLobbyLogsRequest";