+++ sdks/typescript/src/api/resources/cloud/resources/games/resources/namespaces/client/requests/index.ts
@@ -1,0 +1,1 @@
+export { GetGameNamespaceVersionHistoryRequest } from "./GetGameNamespaceVersionHistoryRequest";