+++ sdks/typescript/src/api/resources/cloud/resources/games/resources/namespaces/resources/logs/client/requests/index.ts
@@ -1,0 +1,1 @@
+export { ListNamespaceLobbiesRequest } from "./ListNamespaceLobbiesRequest";