+++ sdks/typescript/src/api/resources/geo/resources/common/types/index.ts
@@ -1,0 +1,2 @@
+export * from "./Coord";
+export * from "./Distance";