+++ sdks/typescript/src/api/resources/group/client/requests/index.ts
@@ -1,0 +1,6 @@
+export { ListSuggestedRequest } from "./ListSuggestedRequest";
+export { SearchRequest } from "./SearchRequest";
+export { GetBansRequest } from "./GetBansRequest";
+export { GetJoinRequestsRequest } from "./GetJoinRequestsRequest";
+export { GetMembersRequest } from "./GetMembersRequest";
+export { GetProfileRequest } from "./GetProfileRequest";