+++ sdks/typescript/src/api/resources/matchmaker/resources/lobbies/client/requests/index.ts
@@ -1,0 +1,5 @@
+export { SetLobbyClosedRequest } from "./SetLobbyClosedRequest";
+export { FindLobbyRequest } from "./FindLobbyRequest";
+export { JoinLobbyRequest } from "./JoinLobbyRequest";
+export { CreateLobbyRequest } from "./CreateLobbyRequest";
+export { ListLobbiesRequest } from "./ListLobbiesRequest";