+++ sdks/typescript/src/serialization/resources/cloud/resources/devices/resources/index.ts
@@ -1,0 +1,2 @@
+export * as links from "./links";
+export * from "./links/types";